backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
SQLAlchemy>=2.0.36
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
python-dotenv>=1.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.12
email-validator>=2.2.0
reportlab>=4.0.9
python-barcode>=0.16.1
python-dateutil>=2.8.2
boto3>=1.28.0
asyncpg>=0.30.0
